dialogs/requirements.txt
openai-whisper
ollama
python-ffmpeg
pyyaml
flask